Create Completing the Sequence
A string word with any combination of the letters "P," "Q," or "R" is supplied to you. Your objective is to figure out how many extra letters must be added to the string to be legitimate. A string is considered legitimate if it can be created by repeatedly concatenating the letter "PQR".